# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BrandOps API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List submissions
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: creatorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, reviewing, approved, revision_requested, paid, rejected]
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit a video for a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Created submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get a submission by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
    patch:
      operationId: updateSubmission
      tags: [submissions]
      summary: Update a submission (approve/reject/request revision)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionUpdate"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, paid, failed]
        - name: creatorId
          in: query
          schema:
            type: integer
        - name: campaignId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment for an approved submission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
...
components:
  schemas:
...
    Campaign:
      type: object
      required:
        - id
        - title
        - description
        - totalBudget
        - payoutPerVideo
        - platform
        - niche
        - status
        - deadline
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        totalBudget:
          type: number
        payoutPerVideo:
          type: number
        platform:
          type: string
          enum: [tiktok, instagram, youtube]
        niche:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed, archived]
        deadline:
          type: string
          format: date-time
        inspirationUrls:
          type: ["string", "null"]
        creatorCount:
          type: integer
        approvedCount:
          type: integer
        pendingCount:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Creator:
      type: object
      required:
        - id
        - name
        - email
        - platform
        - handle
        - niche
        - followerCount
        - engagementRate
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        platform:
          type: string
          enum: [tiktok, instagram, youtube]
        handle:
          type: string
        niche:
          type: string
        followerCount:
          type: integer
        engagementRate:
          type: number
        avatarUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, suspended]
        totalEarnings:
          type: number
        approvedVideos:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Submission:
      type: object
      required:
        - id
        - campaignId
        - creatorId
        - videoUrl
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        creatorId:
          type: integer
        videoUrl:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, reviewing, approved, revision_requested, paid, rejected]
        notes:
          type: ["string", "null"]
        payoutAmount:
          type: ["number", "null"]
        campaign:
          $ref: "#/components/schemas/Campaign"
        creator:
          $ref: "#/components/schemas/Creator"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SubmissionInput:
      type: object
      required:
        - campaignId
        - creatorId
        - videoUrl
      properties:
        campaignId:
          type: integer
        creatorId:
          type: integer
        videoUrl:
          type: string
        thumbnailUrl:
          type: string
    SubmissionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, reviewing, approved, revision_requested, paid, rejected]
        notes:
          type: string
        payoutAmount:
          type: number
    Payment:
      type: object
      required:
        - id
        - submissionId
        - creatorId
        - campaignId
        - amount
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        submissionId:
          type: integer
        creatorId:
          type: integer
        campaignId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, processing, paid, failed]
        paidAt:
          type: ["string", "null"]
          format: date-time
        creator:
          $ref: "#/components/schemas/Creator"
        campaign:
          $ref: "#/components/schemas/Campaign"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaymentInput:
      type: object
      required:
        - submissionId
        - creatorId
        - campaignId
        - amount
      properties:
        submissionId:
          type: integer
        creatorId:
          type: integer
        campaignId:
          type: integer
        amount:
          type: number
    PaymentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, paid, failed]
...